+++ docs/changelog.rst
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+Version 3.5.1
+-------------
+This patch version fixes:
+
+- | the `rerank` not being set bug in :meth:`~weaviate.gql.get.GetBuilder.with_ask`.
+
+- | the bug when using double quotes(`"`) in `question` field in :meth:`~weaviate.gql.get.GetBuilder.with_ask`.
+
+- | the bug where `nearText` filter checks for objects in `moveXXX` clause but never sets it.
+
+
 Version 3.5.0
 -------------
 This minor version contains functionality for the new features introduced in Weaviate ``v1.13.0``.
@@ -20,11 +31,11 @@ This minor version contains functionality for the new features introduced in Wea
 
 Version 3.4.2
 -------------
-| this patch version fixes another bug in :meth:`~weaviate.data.DataObject.exists`.
+| This patch version fixes another bug in :meth:`~weaviate.data.DataObject.exists`.
 
 Version 3.4.1
 -------------
-| this patch version fixes bug in :meth:`~weaviate.data.DataObject.exists`.
+| This patch version fixes bug in :meth:`~weaviate.data.DataObject.exists`.
 
 Version 3.4.0
 -------------